Rinnovata interfaccia grafica pagina film, in particolare pulsanti per l'aggiunta/rimozione del film da una lista; è stato poi aggiunto il pulsante per mostrare la homepage del film, qualora disponibile. Aggiunta activity e relativi fragment per la ricerca degli utenti, con una prima versione che cerca tra gli utenti del database di Cognito. Aggiunta activity "Informazioni", "Impostazioni" e fragment segnalazioni delle notifiche. La descrizione di un film è ora mostrata in inglese nel caso in cui non presente in italiano, e solo se non presente nemmeno in inglese viene mostrato "Non disponibile"; una cosa simile è stata fatta per i titoli. Aggiunta possibilità di ingrandire l'immagine di locandina sulla pagina del film. Modificato logo app.

+++ app/src/main/java/it/unina/ingSw/cineMates20/view/adapter/PendingFriendsRequestsAdapter.java
@@ -14,7 +14,7 @@ import androidx.recyclerview.widget.RecyclerView;
 import java.util.List;
 
 import it.unina.ingSw.cineMates20.R;
-import it.unina.ingSw.cineMates20.controller.NotificationsController;
+import it.unina.ingSw.cineMates20.controller.NotificationController;
 import it.unina.ingSw.cineMates20.model.UserDB;
 
 public class PendingFriendsRequestsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
@@ -86,7 +86,8 @@ public class PendingFriendsRequestsAdapter extends RecyclerView.Adapter<Recycler
         return v -> {
             try{
                 Runnable r = acceptRequestEvent.get(position);
-                deleteItem(position);
+                if(NotificationController.getNotificationControllerInstance().isInternetAvailable())
+                    deleteItem(position);
                 new Thread(r).start();
             } catch(NullPointerException ignore) {}
         };
@@ -97,7 +98,8 @@ public class PendingFriendsRequestsAdapter extends RecyclerView.Adapter<Recycler
         return v -> {
             try{
                 Runnable r = rejectRequestEvent.get(position);
-                deleteItem(position);
+                if(NotificationController.getNotificationControllerInstance().isInternetAvailable())
+                    deleteItem(position);
                 new Thread(r).start();
             } catch(NullPointerException ignore) {}
         };
@@ -113,7 +115,7 @@ public class PendingFriendsRequestsAdapter extends RecyclerView.Adapter<Recycler
         notifyItemRangeChanged(position, users.size());
 
         if(users.size() == 0)
-            NotificationsController.getNotificationControllerInstance().showEmptyNotificationsPage();
+            NotificationController.getNotificationControllerInstance().showEmptyFriendsNotificationPage(true);
     }
 
     @Override
